refactor(fileToBase64)!: unexport with the size-limit helpers
fileToBase64, formatBytes, getMaxFileSize, and fileSizeLimitMessage
had zero external call sites at the v1 sweep (rule 9) — unexport them
from the package root. They stay as internal helpers shared by
useFileUpload, FileUploadHandler, and the editor's media upload
engine.

The two legacy TextEditor extensions that imported fileToBase64
through the public root barrel now import it directly, so removing
the export doesn't break frappe-ui's own build.
